feat: rework and rename  cjsToEsm.cjs to esm.mjs
